fix(desktop): resolve gateway bootstrap review comments
