+++ packages/desktop/src/renderer/hooks/useGatewayDispatcherSetup.ts
@@ -1,16 +1,52 @@
 import { createGatewayDispatcher } from '@clawwork/core';
 import type { ExecApprovalRequest, ModelCatalogEntry, AgentInfo } from '@clawwork/shared';
-import { parseTaskIdFromSessionKey } from '@clawwork/shared';
+import { parseAgentIdFromSessionKey, parseTaskIdFromSessionKey } from '@clawwork/shared';
 import { toast } from 'sonner';
 import { hydrateFromLocal, retrySyncPending, syncFromGateway, syncSessionMessages } from '../lib/session-sync';
 import i18n from '../i18n';
 import { composerBridge, ports, useMessageStore, useTaskStore, useUiStore, useRoomStore } from '../platform';
 import { useApprovalStore } from '../stores/approvalStore';
-import { handlePerformerCandidate, handleSubagentCandidate } from './useEnsembleSync';
 
 export type GatewayDispatcher = ReturnType<typeof createGatewayDispatcher>;
 
 let dispatcher: GatewayDispatcher | null = null;
+
+function handlePerformerCandidate(taskId: string, sessionKey: string): void {
+  if (useRoomStore.getState().lookupTaskIdBySubagentKey(sessionKey)) return;
+
+  const room = useRoomStore.getState().rooms[taskId];
+  if (!room || room.status === 'stopped') return;
+
+  const agentId = parseAgentIdFromSessionKey(sessionKey);
+  useRoomStore.getState().registerPerformerKey(taskId, sessionKey, agentId, agentId);
+  void syncSessionMessages(taskId, sessionKey).catch((err) => {
+    console.error('[performer] syncSessionMessages failed:', err);
+  });
+}
+
+function handleSubagentCandidate(sessionKey: string, gatewayId: string): void {
+  const tasks = useTaskStore.getState().tasks;
+  const ensembleTasks = tasks.filter((task) => task.ensemble && task.gatewayId === gatewayId);
+  if (ensembleTasks.length === 0) return;
+
+  for (const task of ensembleTasks) {
+    const room = useRoomStore.getState().rooms[task.id];
+    if (!room || room.status === 'stopped') continue;
+
+    void useRoomStore
+      .getState()
+      .verifyCandidates(task.id, task.gatewayId)
+      .then(() => {
+        if (useRoomStore.getState().lookupTaskIdBySubagentKey(sessionKey) === task.id) {
+          return syncSessionMessages(task.id, sessionKey);
+        }
+        return undefined;
+      })
+      .catch((err) => {
+        console.error('[subagent] verifyCandidates or syncSessionMessages failed:', err);
+      });
+  }
+}
 
 function handleApprovalRequested(gatewayId: string, payload: unknown): void {
   const approvalReq = payload as ExecApprovalRequest;
@@ -30,7 +66,9 @@ function handleApprovalRequested(gatewayId: string, payload: unknown): void {
           taskId: approvalTaskId,
         });
       })
-      .catch(() => {});
+      .catch((err) => {
+        console.error('[approval] Failed to get settings or send notification:', err);
+      });
   }
 }
